Various Treap updates

+++ src/main/kotlin/datastructures/tree/TreapNode.kt
@@ -4,7 +4,7 @@ interface TreapNode {
 
     var leftChild : TreapNode?
     var rightChild : TreapNode?
-    var value : Int
+    var value : Long
     var priority : Int
     var size : Int
 }
