+++ src/test/kotlin/implementations/TreapNodeImpl.kt
@@ -1,10 +1,9 @@
 package implementations
 
 import datastructures.tree.TreapNode
-import java.util.*
 import kotlin.random.Random
 
-class TreapNodeImpl(override var value: Int) : TreapNode {
+class TreapNodeImpl(override var value: Long) : TreapNode {
     override var leftChild: TreapNode? = null
     override var rightChild: TreapNode? = null
     override var priority: Int = Random.nextInt(Int.MAX_VALUE)
